Template should not fail when value has dollar.
The template uses java regular expression replace to insert the values. The java replaceAll method treats the $ symbol
as a variable for inserting capturing groups. When a dollar symbol is passed in the value part java tries to treat it as
capturing group and fails with an exception.

This fix escapes the $ symbol before it is used