Add overrides for U.get, U.set and U.remove
Accept paths as a List<String>, for it its more usefull than parsing a
complex regex.

One problem of parsing a regex is that you may have a valid separator
as part of a key from your map. For example, if a key is an email, you
could get in trouble for using some email ending in something.com.